Separate Scheme Shard background compaction code and improve logging
